+++ test/package.test.js
@@ -10,10 +10,13 @@ describe('@authnomicon/login', function() {
     
     it('should have assembly metadata', function() {
       expect(json.assembly.namespace).to.equal('org.authnomicon/login');
-      
-      expect(json.assembly.components).to.have.length(4);
-      expect(json.assembly.components).to.include('password/http/service');
-      expect(json.assembly.components).to.include('password/http/scheme');
+      expect(json.assembly.components).to.deep.equal([
+        'http/service',
+        'identifier/http/service',
+        'password/http/service',
+        'password/http/scheme',
+        'signup/http/service'
+      ]);
     });
   });