+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `/signup` endpoint.  Moved from [`@authnomicon/account`](https://github.com/authnomicon/account).
 
 ## [0.0.3] - 2021-12-02
 ### Added